+++ crates/indexer/src/models/processor_statuses.rs
@@ -41,14 +41,14 @@ impl ProcessorStatus {
 
     pub fn from_versions(
         name: &'static str,
-        start_version: i64,
-        end_version: i64,
+        start_version: u64,
+        end_version: u64,
         success: bool,
         details: Option<String>,
     ) -> Vec<Self> {
         let mut status: Vec<Self> = vec![];
         for version in start_version..(end_version + 1) {
-            status.push(Self::new(name, version, success, details.clone()));
+            status.push(Self::new(name, version as i64, success, details.clone()));
         }
         status
     }
